Reworked the entity system
Reworked the entity system to remove inheritance (something that
shouldn't have been there in the first place) and to make an entity a
simple wrapper of component.
Reworked the manager class to use the newly improved Entity.
Introduced a new way to generate ID based on a passed string.
Reworked the Z80 and Ram class to work as components.
Now every component/entity update and refresh function's have the
elapsed time as an argument.
Reworked the TFactory namespace to work with the new entity system