Introduced a Scene class and his manager
Created an IScene class that is an interface to create the game
individual and independent from each other scenes. A scene is a every
portion of th game that has it's own way of loading resources, handling
it's own entities and variables, handling input etc. For example the
main menu and all his attached submenu are just one Scene. The partion
of the game where all the action happens, placing component wiring etc.
is another.
A TSceneManager class is introduced as a way to create and handle the
collecion of scenes present in the game.
Refractored the TGuiManager and the TEntity classes to work with the new
scenes classes.
Refractored the main TAppClass to work with the scene editor