Component system rework
Reworked the component array class. FIrst of all since it's not used
anymore by the manager class but just TEntity, there is no need anymore
to be a templated class.
The TComponentArray in additon doesn't use anymore a shared_ptr to store
is element but an unique_ptr since this class it's the only class who
should add/remove/modify the components in an entity. The components
that use other component now store just a simple pointer to refer to
each other since they don't own the pointed object but they only use to
refer to it.
This overhaul of the class not only improve performance but also make
the class make more sense from a semantics point of view.
Fixed some bugs in the TComponentArray and also added a method to
replace the component that alredy exist with a new component. This
invalidates the pointer to the replaced component used in the other
components. This is something we need to fix in the future.
As a result of all this changes changed the code in the existing
components to use the new TComponentArray class.
Finished the switch to the new ID generation method.