Further GUI work
Removed the dumb TWidgetRenderStruct class. Now the ZSorting happens
when a widget it's added to the vector. Now every widget holds a ZIndex
value that it's used to establish the order of rendering.
Moved the handling of the widgets from IScreenView to a new class called
TWidgetsVector that now is responsible for adding/removing/editing the
widgets.
Layed some ground work work handling the widget state.
Created a scene and relative menu for testing the widget present in the
engine.
Crated some new widgets.