Improvement to TTextComponent
Several improvment to TTextComponent.
- Now this component has is own vertex array and texture from sf::Font.
Thanks to that an entity can both have some sort of graphics (provided
by the TDrawableComponent) and a text!
- Removed the need for 2 sf::VertexArray, this component now just need 1
vertex array that can contain both the outlines vertices and the normal
one. This should also improve performance since we are just doing one
draw call for the entire thing
- Add a lot of customization option for this component! Add the ability
to set the fill color, the outline color for a single character or a
range of character. Also added the ability to customize the
strike-through and underline fill and outline color.

Made some basic change to TDrawableComponent to accomodate the changes
in TTextComponent..