GUI Refactoring
Complete refactoring of the GUI work.
Switch the GUI codebase from a component base system (CBS) to a
traditional inheritance base system. This choice was made since the CBS
proved to be very inappropriate for a simple GUI like the one we need
for this game. Also the GUI for this game is mostly static (doesn't
change often over time) so we can get away with having a refresh/update
function to call at every frame to change the widget state.
Another big change was the removal of the complicated GUI creation
system the required multiple base and derived classes to just display a
single button. The new GUI is based on 2 base classed: TWidget and
TGuiManager.

TWIDGET
This class defines the base properties and methos for every GUI widget.
This class inherites the drawing ability and transformation ability from
SFML classes sf::Drawable and sf::Transformable. In order to create a
new widget we need to inherit from this class and add the appropriate
methods and property. One method that must be overridden is the draw
method used to render the widget.

TGUIMANAGER
This class acts wrapper around a std::vector of shared_ptr of TWidgets
sorted by thei ZIndex value. It has method to add/remove widgets to it's
internal vector used to keep track of all the widgets present in a
scene. The widgets at the moment of being added to the vector can
specify It also defines basic methods to search and retrieve widgets by
WidgetID or WidgetName. The widgets at the moment of being added to the
vector can also specify a ZIndex value to specify the order of
rendering, by default this value it's 0.
This class also defines some methods to iterate over the widgets.

The IMenu class it's a pure virtual class used in tandem with TContainer
to define a menu which is basically a collection of widgets with one
common parent Widget. This class definies just a single pure virtual
function init. This function it's used by derived class to setup the
list of widgets the form the menu.

Basic widget are defined: TButton, TContainer, TImage, TStaticText.