Small bug-fixes commit
This commit just fixes some minor bugs in the bus/wire component. It
also makes the TZ80Component more indipendent from the RAM component
(now it will "work" without a ram attached to it). It also modified the
code of TChipComponent to allow it to be moved around while we are
trying to place the chip and check for collision with other chips
already placed on the logic board.