Changes to TChipComponent and TPinComponent
Big changes to the behaviour of TChipComponent. Before this chip was
pesonally responsable of the rendering of the variuos elettrical
component and the pin of such component. Both this functionality are now
removed from TChipComponent.
The rendering capability are now taken care of by specialized classes
that implement all the visual aspect of an elettronic component such as
TPackageComponent that implements the rendering of various semiconductor
package or TLedComponent that implements a basic LED component.
All the methods that regarded the PIN were moved to TPinComponent.
Two new classes are introduced, the previous mentioned TPackageComponent
and TLedComponent. The first implement a basic DIP package rendering for
the future IC (for now works great with the z80 and ram component) and
nothing more. The second implement a basic LED system that renders a
basic LED that "lights up" when the ANODE pin is set to high. The
component it's a very primordial phase and lacks of a lot of real
methods and function to make it work like a real elettronic component
with is own resistence, voltage and current.

All the other changes in the other classed in the commits are a
reflection of all the changes mentioned above.