Creation of a TPowerComponent and other enanchement
A TPowerComponent has been created that implements both a power source
(like a battery) and  a ground point. When this component it's created
we need to specify if the component is gonna act as a power source or
ground, this cannot be changed lated and will affect the behavior of the
component during it's lifetime. Upon initialization the component load
the proper texture, creates a single POWER pin in his TPinComponent, and
adds a graphical PIN to the vertex array. The TPackageComponent was
changed to work in conjuction with this newly created TPowerComponent.
Now the TPackageComponent feature a variable that holds whatever the
TPackageComponent it's powered on or not. This variable can be retrive
via the isPoweredOn() function and used by the other component should
execute eventual refresh/update cycle or other operation that have to be
executed only if the chip is powered on.

The TPinComponent was removed from the tcomponents namespace since now
it became a "graphic" component while before it was a pure logicical
component. Also added a way to search and get a pin in his internal
vector by pin name. A static TPin variable was created that implement
the NotFound pin that can be used by the getPin(...) functions if the
pin we are looking for can't be found.

The TLedComponent now lights up only if both is pin are connected to
something and the anode is positive while the cathode is negative. Still
not a real check of the eletrical properties of the component.

We also updated the TLogicBoardComponent to also be able to remove a
chip, wire, bus from his internal vectors.

We also changed the code in TNewGameScene to avoid creating 2 z80 or Ram
entity. We also updated the code so everytime we are placing a new
chip/wire/bus check if we were already trying to place an entity, if so
remove the temp entity from the TLogicBoardComponent and the vector of
entity used by the scene.

All the other changes in the other classed in the commits are a
reflection of all the changes mentioned above.