GUI Update
The Widget classed was updated to support basic event handling
(attaching/detaching/firing an event) and now this class also take
personally care of handling his internal state. As a result of the
latter the TStataManager class was removed.
The event handling is very basic right now. It only defines a simple
mouse click event that get triggered (aka fired) when the user click on
an enabled widget.
A update method was also included that take care of updating the widget
look. All derived classes that implement a custom look based on the
widget state need to override this function.
A TMenuButton class was introduced to implement the button present in
the main menus.
Also a TComplexText class was introduced, this class is a copy of
TTextComponent, stripped of is reliance on the ECS and made more similar
to a SFML class.