Big changes to several classes
This commit features a lot of changes to the core functionality as well
as the introduction of several new classes.

GUI Changes
The list of GUI events is expanded and partially implemented.
The update() implementation of TButton is changed to work with the state
management system introduced in the previus commit.

A new GUI widget called TImageButton is introduced. This class is very
similar to TButton (from which we inherit) but it introduce the ability
to add an icon to the button. At the moment the functionality of the
class it's very limited, the idea is to add in the future several
methods to customize the way this image is displayed within the button.

A new GUI widget called TRenderCanvas is introduced. This class inherit
from TWidget and replaces the TRenderSurface class (which is now
depracted) and expand his functionality. This widget acts as a canvas in
which graphical entity can be drawn into. Kinda act as a
sf::RenderTexture but adds other functionality such as the ability to
resize/move/zoom it.

TEVENTCOMPONENT AND TSTATECOMPONENT
Two new classes are intoduced: TEventComponent and TStateComponent. The
former adds the event handling for the Entity and his component and
stores the position of the mouse within the canvas on which the entity
is rendered. This class his a copycat of the event handling portion of
TWidget.
The ladder implements a basic TStateManager that stores the current
status of the Entity. Now all the other components can base their looks
and/or logic based on the Entity state.

TLOGICBOARDCOMPONENT
This class will keep record of which kind of insertion method we are
using and a string that holds the last added chip.

TMANAGER
The entity manager class now will handle the event dispatching via the
processEvents() function.

All the other changes in the other classed in the commits are a
reflection of all the changes mentioned above.