GUI update and fixes
A new class TDialogWindow is added that implements a simple popup modal
window. The class at the moment of creation it's added to the gui
manager. Whn the window is closed we can choose if the widget should be
removed from the manager or not. The user can also personalize the title
caption and message.

Fixed the rendering of widgets with a setted parent. Now when calling
the draw function the widget will use the right transform. Also changed
the method to check for visibility, now return true if both the widget
itself and all his parent widget are visible.
Fixed the rendering of toggleable widget. Now the widget will
appropriately update his status when we click on it. It will change to
Selected if it's in his normal state, and go back to Normal if it's
selected.

All the other changes in the other class in the commits are a reflection
of all the changes mentioned above.

Several bugfixes and improvments in the other class in the commit