Introduced debugger tools classes
Introduced a TCodeEditor class that will act as our text editor for the
z80 programs. For now it act as a read only editor that can ready a
program loaded in memory.
Introduced also a TDebugWindow class that acts as a floating window that
output a debug value. For now we use to visualize the value of the z80
registers and flags.

All the other changes in the other class in the commits are a reflection
of all the changes mentioned above.

Several bugfixes and improvments in the other class in the commit