Made possible to create a connection between a wire and a chip
Changes in code to make possible to connect a PIN from an existing wire
as opposed as from just  another PIN. Now the wire stores a vector of
PIN that are connected to it.

All the other changes in the other class in the commits are a reflection
of all the changes mentioned above.

Several bugfixes and improvments in the other class in the commit