Introduced XML file support and support for loading/saving a board
In this commit we introduced support for loading/saving and handling
.xml file through the TXMLFile and TXMLNode class. The latter rappresent
a node in a .xml file. It holds the name, the attribute, the node's
content and his children if it has any. This class allows basic
manipulation of all the characteristics mentioned above plus methods for
retrieving them. The former allows to read and save a valid XML file.
When loading/saving we have to supply a XMLNode variable that will hold
the loaded data in case of loading or has the data to save if we are
saving. The saving/loading it's effectuated in the TLogicBoardComponent.

We also added a variable that holds the chip name in TChipComponent. We
use this variable to link a PIN to the chip that holds it.

Made a change in the TPin code comparison operator. Now we can compare a
chip with another TStatus or another TPin. Also now we keep track of all
the pins in the logic board in a static vector. A pin it's added to this
vector in his constructor and removed it in his destructor. The pin in
this vector can be retrieved by 2 methods. The first use the PinID to
look in the vector, the second use the name of the PIN and his ower (a
TChipComponent name).

All the other changes in the other class in the commits are a reflection
of all the changes mentioned above.

Several bugfixes and improvments in the other class in the commit.