Changes in the logic board system
The TLogicBoardComponent system and all his functionality have been
completly reworked and rewritted.
We introduced a class TBoard that implements almost all the old methods
and functionality. This class keeps track of all the wire/bus/chip
placed in the logic board by storing a pointer to their component in 3
separated vectors. This class also feature several templated methods to
access/insert/remove the components stored in the vectors.
The old TLogicBoardComponent now only stores a pointer to the TBoard
class and provide methods for setting/getting the TBoard class.

Some fixes in the gui to when it comes to toggleable buttons.

Some changes in the TCodeEditor class to increase performance by
removing the number of object to draw and to keep track of whick
instruction we are currently executing.

All the other changes in the other class in the commits are a reflection
of all the changes mentioned above.

Several bugfixes and improvments in the other class in the commit.