Added support for breakpoints and step-by-step execution
Introduced a new namespace for debugging tools called "debugger".
A new class TBreakpoint is introduced that implements a visual and logic
breakpoint in the program execution. The breakpoint stores a sf::Sprite
object to visually render the breakpoint, and the position in memory of
the memory address and the codeline at which the program should be
stopped.

Support for step-by-step execution and normal contiguous execution.

All the other changes in the other class in the commits are a reflection
of all the changes mentioned above.

Several bugfixes and improvments in the other class in the commit.