Added support for removing an inserted wire
A new tool to remove components inserted in the logic board is added.
For now we can only remove wire.
The TWireComponent class was reworked. Now the wire stores not only the
pin connected to the wire but also other wire connected to the wire.
When a wire it's connected we also connect all the pin together. When a
wire it's detach we detach all the connected pin.

Save changes in the TBoard class to the save/load methods to accomodate
for the changes in TWireComponent