Initial support for removing chip and pin connections fixes
Added the ability to remove an inserted chip from the logic board and
entity mamanger. This functionality is a little big bugged. While the
chip is removed correctly, the detachment of the his pin to theirs
connected pins it's a little bit problematic due to how pin connections
is implemented.

Some bug in the pin/wire connections were fixed and some potential fixes
for future bug-fixes were identified but non timplemented yet.

The TPackageComponent now it's responsible for the rendering of all kind
of chip, not only DIP package but also LED, power connectors etc. This
component now also feature a more streamlined way of creating labels and
pin to work with every type of chip.

All the other changes in the other class in the commits are a reflection
of all the changes mentioned above.

Several bugfixes and improvments in the other class in the commit.